requirements.txt
deepagents>=0.4.1
langchain-openai>=1.1.0
langchain-anthropic>=1.3.0
langgraph-checkpoint-sqlite>=3.0.0
tavily-python>=0.7.0
python-dotenv>=1.0.0
alibabacloud_bailian20231229>=1.0.0
alibabacloud_tea_openapi>=0.3.0
alibabacloud_tea_util>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
jieba>=0.42.1
numpy>=1.24.0
openai>=1.0.0
requests>=2.31.0
pdfplumber>=0.10.0
python-docx>=1.0.0
openpyxl>=3.1.0
python-multipart>=0.0.7
